Graphics API: draw line using floats

+++ vtm-android/src/org/oscim/android/canvas/AndroidCanvas.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016 devemux86
+ * Copyright 2016-2017 devemux86
  * Copyright 2017 nebular
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
@@ -68,7 +68,7 @@ public class AndroidCanvas implements Canvas {
     }
 
     @Override
-    public void drawLine(int x1, int y1, int x2, int y2, Paint paint) {
+    public void drawLine(float x1, float y1, float x2, float y2, Paint paint) {
         canvas.drawLine(x1, y1, x2, y2, ((AndroidPaint) paint).mPaint);
     }
